added headers and sources for stats collection and analysis

+++ include/statsdb/game_data.hpp
@@ -82,7 +82,7 @@ struct SpecialTeams {
     };
     int goals;
     int attempts;
-    double get_efficiency(Type type) const;
+    float get_efficiency(Type type) const;
 };
 
 using IntResult = TeamsValueHolder<int>;
